Compartimentation des Classes par type (Weapon, Player, Monster, Fonctions)
